不要なコードの削除

+++ src/main/java/com/example1/demo3/controller/api/StockHistoryInApiController.java
@@ -21,7 +21,7 @@ public class StockHistoryInApiController {
     }
 
     @GetMapping("/{productId}")
-    public List<StockHistoryInDto> getInHistory(@PathVariable Long productId) {
+    public List<StockHistoryInDto> getInHistory(@PathVariable Integer productId) {
         return service.getInHistory(productId);
     }
 }
